Housekeeping of gridarchitect. Removed old and unnecessary code, that was never used in our Scenarios.
Note, the removed code was mainly the Thesis from Luc, many years ago.
