+++ core/gridarchitect/src/main/java/akka/BlankActorBehavior.java
@@ -9,9 +9,6 @@
 
 package akka;
 
-import java.util.LinkedList;
-
-import akka.advancedMessages.ErrorAnswerContent;
 import akka.basicMessages.AnswerContent;
 import akka.basicMessages.RequestContent;
 import behavior.BehaviorModel;
@@ -40,11 +37,6 @@ public class BlankActorBehavior extends BehaviorModel{
 	@Override
 	public RequestContent returnRequestContentToSend() {
 		return null;
-	}
-
-
-	@Override
-	public void handleError(LinkedList<ErrorAnswerContent> errors) {
 	}
 
 }
