--- core/gridarchitect/src/main/java/akka/advancedMessages/ForwardActorFailureAnswerContent.java
@@ -1,21 +1,0 @@
-/*
- * Copyright (c) 2011-2015, fortiss GmbH.
- * Licensed under the Apache License, Version 2.0.
- *
- * Use, modification and distribution are subject to the terms specified
- * in the accompanying license file LICENSE.txt located at the root directory
- * of this software distribution.
- */
-
-package akka.advancedMessages;
-
-import akka.basicMessages.AnswerContent;
-/**
- * special answer that is the answer to a ForwardActorFailureRequestContent
- * @author Luc
- *
- */
-public class ForwardActorFailureAnswerContent implements AnswerContent {
-
-
-}
