+++ core/gridarchitect/src/main/java/akka/basicMessages/BasicAnswer.java
@@ -24,8 +24,6 @@ public class BasicAnswer {
     
 	public final int timeStep;
     public final List<ActorRef> upstreamActorTrace;
-    public final boolean sane;
-    public final boolean overrideReportToParent;
     public final String senderPath;
     public final AnswerContent answerContent;
 
@@ -41,15 +39,11 @@ public class BasicAnswer {
      */
     public BasicAnswer(int timeStep, 
     		List<ActorRef> upstreamActorTrace, 
-    		boolean sane,
-    		boolean overrideReportToParent, 
     		String senderPath, 
     		AnswerContent answerContent) {
     	
         this.timeStep = timeStep;
         this.upstreamActorTrace = upstreamActorTrace;
-        this.sane = sane;
-        this.overrideReportToParent = overrideReportToParent;
         this.senderPath = senderPath;
         this.answerContent = answerContent;
     }
@@ -59,8 +53,6 @@ public class BasicAnswer {
         return "BasicAnswer{" +
                 "timeStep=" + timeStep +
                 ", actorTrace=" + upstreamActorTrace +
-                ", sane=" + sane +
-                ", overrideReportToParent=" + overrideReportToParent +
                 ", senderPath='" + senderPath + '\'' +
                 ", behaviorMessage='" + answerContent + '\'' +
                 '}';
