+++ core/gridarchitect/src/main/java/behavior/InactiveBehaviorModel.java
@@ -9,9 +9,6 @@
 
 package behavior;
 
-import java.util.LinkedList;
-
-import akka.advancedMessages.ErrorAnswerContent;
 import akka.basicMessages.AnswerContent;
 import akka.basicMessages.RequestContent;
 
@@ -33,11 +30,6 @@ public class InactiveBehaviorModel extends BehaviorModel {
 	@Override
 	public RequestContent returnRequestContentToSend() {
 		return null;
-	}
-
-	@Override
-	public void handleError(LinkedList<ErrorAnswerContent> errors) {
-		
 	}
 
 }
