--- core/gridarchitect/src/main/java/faultStrategy/requests/HistoryRequestContent.java
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) 2011-2015, fortiss GmbH.
- * Licensed under the Apache License, Version 2.0.
- *
- * Use, modification and distribution are subject to the terms specified
- * in the accompanying license file LICENSE.txt located at the root directory
- * of this software distribution.
- */
-
-package faultStrategy.requests;
-
-import akka.basicMessages.RequestContent;
-
-public abstract class HistoryRequestContent implements RequestContent{
-
-}
