+++ projects/memapCore/src/main/java/memap/components/prototypes/MEMAPCoordination.java
@@ -1,11 +1,9 @@
 package memap.components.prototypes;
 
 import java.util.ArrayList;
-import java.util.LinkedList;
 
 import com.google.gson.Gson;
 
-import akka.advancedMessages.ErrorAnswerContent;
 import akka.basicMessages.AnswerContent;
 import akka.basicMessages.BasicAnswer;
 import akka.basicMessages.RequestContent;
@@ -66,10 +64,6 @@ public class MEMAPCoordination extends BehaviorModel implements CurrentTimeStepS
 		nStepsMPC = topologyConfig.getNrStepsMPC();
 		lpSolHandler = new SolutionHandler(nStepsMPC, topologyConfig);
 		milpSolHandler = new SolutionHandler(nStepsMPC, topologyConfig);
-	}
-
-	@Override
-	public void handleError(LinkedList<ErrorAnswerContent> errors) {
 	}
 
 	@Override
